pyproject.toml
[tool.poetry]
name = "bitcoin-sentiment-mlops"
version = "1.0.0"
description = "Bitcoin Sentiment Analysis MLOps Showcase - Production-grade ML system"
authors = ["Ahmed Ismail Khalid <ahmedik95316@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework & API
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"

# Database
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"

# Redis & Caching
redis = "^5.0.1"
hiredis = "^2.3.2"

# Data Processing
pandas = "^2.1.4"
numpy = "^1.26.3"
scipy = "^1.11.4"

# Data Validation
pandera = "^0.18.0"

# Web Scraping & APIs
requests = "^2.31.0"
aiohttp = "^3.9.1"
beautifulsoup4 = "^4.12.3"
feedparser = "^6.0.11"
praw = "^7.7.1"

# ML & NLP
scikit-learn = "^1.4.0"
transformers = "^4.36.2"
torch = "^2.1.2"
vadersentiment = "^3.3.2"

# Monitoring & Logging
prometheus-client = "^0.19.0"
python-json-logger = "^2.0.7"

# Utilities
python-dotenv = "^1.0.0"
tenacity = "^8.2.3"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
# Code Quality
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pylint = "^3.0.3"

# Testing
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.3"
pytest-mock = "^3.12.0"

# Development Tools
ipython = "^8.19.0"
jupyter = "^1.0.0"

[tool.poetry.scripts]
setup-env = "scripts.development.setup_environment:main"
create-tables = "scripts.development.create_tables:main"
collect-data = "scripts.data_collection.collect_all_data:main"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"